hdfs.py refactoring: Move out format
(Ok, I cheated a bit, the tests will fail for this commit)
