+++ luigi/contrib/hdfs/__init__.py
@@ -43,6 +43,18 @@ from luigi.contrib.hdfs.config import load_hadoop_cmd
 from luigi.contrib.hdfs.config import tmppath
 from luigi.contrib.hdfs import config as hdfs_config
 
+# Here follows a series of deprecated imports that have been moved out there own files.
+from luigi.contrib.hdfs import format as hdfs_format
+
+HdfsReadPipe = hdfs_format.HdfsReadPipe
+HdfsAtomicWritePipe = hdfs_format.HdfsAtomicWritePipe
+HdfsAtomicWriteDirPipe = hdfs_format.HdfsAtomicWriteDirPipe
+PlainFormat = hdfs_format.PlainFormat
+PlainDirFormat = hdfs_format.PlainDirFormat
+Plain = hdfs_format.Plain
+PlainDir = hdfs_format.PlainDir
+CompatibleHdfsFormat = hdfs_format.CompatibleHdfsFormat
+
 logger = logging.getLogger('luigi-interface')
 
 
@@ -590,130 +602,6 @@ rename = client.rename
 remove = client.remove
 mkdir = client.mkdir
 listdir = client.listdir
-
-
-class HdfsReadPipe(luigi.format.InputPipeProcessWrapper):
-
-    def __init__(self, path):
-        super(HdfsReadPipe, self).__init__(load_hadoop_cmd() + ['fs', '-cat', path])
-
-
-class HdfsAtomicWritePipe(luigi.format.OutputPipeProcessWrapper):
-    """
-    File like object for writing to HDFS
-
-    The referenced file is first written to a temporary location and then
-    renamed to final location on close(). If close() isn't called
-    the temporary file will be cleaned up when this object is
-    garbage collected
-
-    TODO: if this is buggy, change it so it first writes to a
-    local temporary file and then uploads it on completion
-    """
-
-    def __init__(self, path):
-        self.path = path
-        self.tmppath = tmppath(self.path)
-        parent_dir = os.path.dirname(self.tmppath)
-        mkdir(parent_dir, parents=True, raise_if_exists=False)
-        super(HdfsAtomicWritePipe, self).__init__(load_hadoop_cmd() + ['fs', '-put', '-', self.tmppath])
-
-    def abort(self):
-        logger.info("Aborting %s('%s'). Removing temporary file '%s'",
-                    self.__class__.__name__, self.path, self.tmppath)
-        super(HdfsAtomicWritePipe, self).abort()
-        remove(self.tmppath)
-
-    def close(self):
-        super(HdfsAtomicWritePipe, self).close()
-        rename(self.tmppath, self.path)
-
-
-class HdfsAtomicWriteDirPipe(luigi.format.OutputPipeProcessWrapper):
-    """
-    Writes a data<data_extension> file to a directory at <path>.
-    """
-
-    def __init__(self, path, data_extension=""):
-        self.path = path
-        self.tmppath = tmppath(self.path)
-        self.datapath = self.tmppath + ("/data%s" % data_extension)
-        super(HdfsAtomicWriteDirPipe, self).__init__(load_hadoop_cmd() + ['fs', '-put', '-', self.datapath])
-
-    def abort(self):
-        logger.info("Aborting %s('%s'). Removing temporary dir '%s'",
-                    self.__class__.__name__, self.path, self.tmppath)
-        super(HdfsAtomicWriteDirPipe, self).abort()
-        remove(self.tmppath)
-
-    def close(self):
-        super(HdfsAtomicWriteDirPipe, self).close()
-        rename(self.tmppath, self.path)
-
-
-class PlainFormat(luigi.format.Format):
-
-    input = 'bytes'
-    output = 'hdfs'
-
-    def hdfs_writer(self, path):
-        return self.pipe_writer(path)
-
-    def hdfs_reader(self, path):
-        return self.pipe_reader(path)
-
-    def pipe_reader(self, path):
-        return HdfsReadPipe(path)
-
-    def pipe_writer(self, output_pipe):
-        return HdfsAtomicWritePipe(output_pipe)
-
-
-class PlainDirFormat(luigi.format.Format):
-
-    input = 'bytes'
-    output = 'hdfs'
-
-    def hdfs_writer(self, path):
-        return self.pipe_writer(path)
-
-    def hdfs_reader(self, path):
-        return self.pipe_reader(path)
-
-    def pipe_reader(self, path):
-        # exclude underscore-prefixedfiles/folders (created by MapReduce)
-        return HdfsReadPipe("%s/[^_]*" % path)
-
-    def pipe_writer(self, path):
-        return HdfsAtomicWriteDirPipe(path)
-
-
-Plain = PlainFormat()
-PlainDir = PlainDirFormat()
-
-
-class CompatibleHdfsFormat(luigi.format.Format):
-
-    output = 'hdfs'
-
-    def __init__(self, writer, reader, input=None):
-        if input is not None:
-            self.input = input
-
-        self.reader = reader
-        self.writer = writer
-
-    def pipe_writer(self, output):
-        return self.writer(output)
-
-    def pipe_reader(self, input):
-        return self.reader(input)
-
-    def hdfs_writer(self, output):
-        return self.writer(output)
-
-    def hdfs_reader(self, input):
-        return self.reader(input)
 
 
 class HdfsTarget(FileSystemTarget):
